refactor: move getParentOfNodeType to utils
